Webhook app clean up. (#11756)
