+++ saleor/graphql/webhook/mutations/__init__.py
@@ -1,0 +1,15 @@
+from .event_delivery_retry import EventDeliveryRetry
+from .webhook_create import WebhookCreate
+from .webhook_delete import WebhookDelete
+from .webhook_dry_run import WebhookDryRun
+from .webhook_trigger import WebhookTrigger
+from .webhook_update import WebhookUpdate
+
+__all__ = [
+    "EventDeliveryRetry",
+    "WebhookCreate",
+    "WebhookDelete",
+    "WebhookDryRun",
+    "WebhookTrigger",
+    "WebhookUpdate",
+]
